Remove reducers columns tests
